refactor:💡 schedule response dto API별 폴더링
