+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/getLatestScheduleByTeacher/DateAndDayOfWeek.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.schedule.dto.response;
+package gwasuwonshot.tutice.schedule.dto.response.getLatestScheduleByTeacher;
 
 import lombok.*;
