+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/getTemporarySchedule/GetTemporaryScheduleResponse.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.schedule.dto.response;
+package gwasuwonshot.tutice.schedule.dto.response.getTemporarySchedule;
 
 import lombok.*;
