+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/getTodayScheduleByTeacher/GetTodayScheduleByTeacherResponse.java
@@ -1,6 +1,6 @@
-package gwasuwonshot.tutice.schedule.dto.response;
+package gwasuwonshot.tutice.schedule.dto.response.getTodayScheduleByTeacher;
 
-import gwasuwonshot.tutice.schedule.dto.ScheduleResponse;
+import gwasuwonshot.tutice.schedule.dto.response.ScheduleResponse;
 import gwasuwonshot.tutice.schedule.entity.Schedule;
 import lombok.*;
