+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/getTodayScheduleByTeacher/TodayScheduleDetailByTeacher.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.schedule.dto.response;
+package gwasuwonshot.tutice.schedule.dto.response.getTodayScheduleByTeacher;
 
 import lombok.*;
