+++ src/main/java/gwasuwonshot/tutice/schedule/service/ScheduleService.java
@@ -12,7 +12,16 @@ import gwasuwonshot.tutice.lesson.repository.LessonRepository;
 import gwasuwonshot.tutice.schedule.dto.request.GetTemporaryScheduleRequest;
 import gwasuwonshot.tutice.schedule.dto.request.UpdateScheduleAttendanceRequest;
 import gwasuwonshot.tutice.schedule.dto.request.UpdateScheduleRequest;
-import gwasuwonshot.tutice.schedule.dto.response.*;
+import gwasuwonshot.tutice.schedule.dto.response.getLatestScheduleByTeacher.GetLatestScheduleByTeacherResponse;
+import gwasuwonshot.tutice.schedule.dto.response.getMissingAttendanceScheduleByTeacher.GetMissingAttendanceScheduleResponse;
+import gwasuwonshot.tutice.schedule.dto.response.getMissingAttendanceScheduleByTeacher.MissingScheduleByDate;
+import gwasuwonshot.tutice.schedule.dto.response.getScheduleByUser.GetScheduleByUserResponse;
+import gwasuwonshot.tutice.schedule.dto.response.getScheduleByUser.ScheduleByDate;
+import gwasuwonshot.tutice.schedule.dto.response.getTemporarySchedule.GetTemporaryScheduleResponse;
+import gwasuwonshot.tutice.schedule.dto.response.getTemporarySchedule.TemporarySchedule;
+import gwasuwonshot.tutice.schedule.dto.response.getTodayScheduleByParents.GetTodayScheduleByParentsResponse;
+import gwasuwonshot.tutice.schedule.dto.response.getTodayScheduleByTeacher.GetTodayScheduleByTeacherResponse;
+import gwasuwonshot.tutice.schedule.dto.response.updateScheduleAttendance.UpdateScheduleAttendanceResponse;
 import gwasuwonshot.tutice.schedule.entity.Schedule;
 import gwasuwonshot.tutice.schedule.entity.ScheduleStatus;
 import gwasuwonshot.tutice.schedule.exception.*;
